+++ src/components/inputs/index.ts
@@ -1,0 +1,3 @@
+export { CPTAddressBlock } from './CPTAddressBlock';
+export type { CPTAddressBlockProps, AddressData } from './CPTAddressBlock';
+